Re-organized characters files into separate pkg
