+++ app/src/main/res/layout/fragment_characters.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CharactersFragment">
+    tools:context=".characters.CharactersFragment">
 
     <ProgressBar
         android:id="@+id/progressBar"
